Fix(Pos): 사업자 인증 통합
